mobius-server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
redis==5.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
langchain==0.3.9
langgraph==0.2.53
litellm==1.52.14
httpx==0.27.2
apscheduler==3.10.4
cryptography==43.0.3
pytest==8.3.3
pytest-asyncio==0.24.0
email-validator==2.1.1
celery[redis]>=5.3.0
